Added protection against IDOR vulnerability (Insecure Direct Object Reference)

+++ routes/render/users.route.js
@@ -13,42 +13,51 @@ router.route('/')
 router.route('/:id')
   .get((req, res) => {
     const { id } = req.params;
-    const { user } = req.session;
+    const { user } = req;
 
     res.renderComponent(Edit, { user, id, title: 'Edit your profile' });
   })
   .put((req, res) => {
     const { id } = req.params;
 
-    User.update(req.body, { where: { id }, returning: true, raw: true })
-      .then((updatedUser) => {
-        const [, [update]] = updatedUser;
-        // req.session.user = updatedUser[1][0];
-        req.session.user = update;
-        res.json({ updated: true, url: '/profile' });
-      })
-      .catch((error) => res.status(500).json({ error: error.message }));
+    if (req.session.userId === Number(id)) {
+      User.update(req.body, { where: { id }, returning: true, raw: true })
+        .then((updatedUser) => {
+          const [, [update]] = updatedUser;
+          // req.session.user = updatedUser[1][0];
+          req.session.userId = update.id;
+          res.json({ updated: true, url: '/profile' });
+        })
+        .catch((error) => res.status(500).json({ error: error.message }));
+    } else {
+      res.status(403).json({ message: 'Access Denied' });
+    }
   })
   .delete((req, res) => {
     const { id } = req.params;
 
-    User.destroy({ where: { id } })
-      .then((deletedUser) => {
-        if (deletedUser) {
-          // удаление сессии на сервере
-          req.session.destroy();
+    // проверка сессии с ID пользователя для удаления текущего пользователя (IDOR)
+    if (req.session.userId === Number(id)) {
+      User.destroy({ where: { id } })
+        .then((deletedUser) => {
+          if (deletedUser) {
+            // удаление сессии на сервере
+            req.session.destroy();
 
-          // серверное удаление куки по имени
-          res.clearCookie('user_uid');
+            // серверное удаление куки по имени
+            res.clearCookie('user_uid');
 
-          // формирование ответа при успешном удалении
-          res.json({ delete: true, url: '/' });
-        } else {
-          // формирование ответа при некоректном удалении
-          res.status(404).json({ delete: false });
-        }
-      })
-      .catch((error) => res.status(500).json({ error: error.message }));
+            // формирование ответа при успешном удалении
+            res.json({ delete: true, url: '/' });
+          } else {
+            // формирование ответа при некоректном удалении
+            res.status(404).json({ delete: false });
+          }
+        })
+        .catch((error) => res.status(500).json({ error: error.message }));
+    } else {
+      res.status(403).json({ message: 'Access Denied' });
+    }
   });
 
 module.exports = router;
